requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.2
Django==4.0.3
django-colorfield==0.6.3
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
jwt==1.3.1
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
